# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Agorich Pharma API
  description: |
    Comprehensive API documentation for Agorich Pharma's hyperlocal pharma intelligence operating system.

    ## Overview
    This API provides endpoints for managing the pharmaceutical distribution ecosystem including:
    - AI-powered collection calls via VAPI
    - Inventory management and alerts
    - Invoice processing and approval workflows
    - Real-time metrics and analytics
    - Emergency controls and system management

    ## Authentication
    - **User Authentication**: Supabase Auth with JWT tokens (Cookie-based for web clients)
    - **Agent API Key**: `x-agent-api-key` header for backend-to-backend communication
    - **Admin API Key**: `x-admin-api-key` header for admin operations

    ## Rate Limiting
    - Default: 100 requests per minute per IP
    - AI endpoints: 20 requests per minute
    - Webhook endpoints: 200 requests per minute
  version: 1.0.0
  contact:
    name: Agorich Pharma Tech Team
    email: tech@agorichpharma.com
  license:
    name: Proprietary
servers:
  - url: https://api.agorichpharma.com
    description: Production server
  - url: https://staging-api.agorichpharma.com
    description: Staging server
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/admin/accounts-receivable:
    get:
      tags:
        - Admin
      summary: Get accounts receivable
      description: Returns all invoices with outstanding balance due.
      operationId: getAccountsReceivable
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, partially_paid, overdue, all]
        - name: sort_by
          in: query
          schema:
            type: string
            default: due_date
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: AR data retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsReceivableResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/admin/notifications:
    get:
      tags:
        - Admin
      summary: Get system notifications
      description: Retrieves system notifications for admin dashboard.
      operationId: getNotifications
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Notifications retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationsResponse'
    post:
      tags:
        - Admin
      summary: Create notification
      description: Creates a new system notification.
      operationId: createNotification
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationCreate'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-agent-api-key
      description: API key for agent-to-agent communication
    AdminApiKeyAuth:
      type: apiKey
      in: header
      name: x-admin-api-key
      description: Admin API key for administrative operations
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Supabase Auth
    WebhookSecretAuth:
      type: apiKey
      in: header
      name: x-vapi-webhook-secret
      description: Webhook verification secret
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message description"
      required:
        - success
        - error
...
    AccountsReceivableInvoice:
      type: object
      properties:
        id:
          type: string
        invoice_number:
          type: string
        customer_id:
          type: string
        customer_name:
          type: string
        customer_business_name:
          type: string
        customer_phone:
          type: string
        grand_total:
          type: number
        advance_paid:
          type: number
        balance_due:
          type: number
        payment_status:
          type: string
        gst_type:
          type: string
        place_of_supply:
          type: string
        invoice_date:
          type: string
          format: date
        due_date:
          type: string
          format: date
        status:
          type: string
        days_overdue:
          type: integer
        created_at:
          type: string
          format: date-time
    AccountsReceivableResponse:
      type: object
      properties:
        success:
          type: boolean
        invoices:
          type: array
          items:
            $ref: '#/components/schemas/AccountsReceivableInvoice'
        summary:
          type: object
          properties:
            totalOutstanding:
              type: number
            totalAdvanceCollected:
              type: number
            invoiceCount:
              type: integer
...
    Notification:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        title:
          type: string
        message:
          type: string
        read:
          type: boolean
        created_at:
          type: string
          format: date-time
    NotificationsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        unreadCount:
          type: integer
    NotificationCreate:
      type: object
      required:
        - type
        - title
        - message
      properties:
        type:
          type: string
        title:
          type: string
        message:
          type: string
        user_id:
          type: string
...